refactor(snapshot): move snapshot policy utilities (#2135)
