+++ src/commands/interaction/runtime/gestures.ts
@@ -26,7 +26,7 @@ import {
   type ScrollEdge,
   type ScrollEdgeState,
   type ScrollEdgeTarget,
-} from '../../../utils/scroll-edge-state.ts';
+} from '../../../snapshot/scroll-edge-state.ts';
 import { successText } from '@agent-device/kernel/success-text';
 import { requireIntInRange } from '../../../utils/validation.ts';
 import { toBackendContext } from '../../runtime-common.ts';
