+++ src/snapshot/__tests__/scroll-edge-state-broad-selection.test.ts
