+++ src/snapshot/__tests__/scroll-edge-state-pass-orchestration.test.ts
@@ -4,7 +4,7 @@ import {
   captureScrollEdgeState,
   formatScrollEdgeMessage,
   runScrollEdgePasses,
-} from './scroll-edge-state.ts';
+} from '../scroll-edge-state.ts';
 import { AppError } from '@agent-device/kernel/errors';
 import { captureThrows, scrollSnapshot, windowRoot } from './scroll-edge-state-fixtures.ts';
