+++ src/snapshot/__tests__/scroll-edge-state-target-resolution.test.ts
