+++ src/snapshot/__tests__/snapshot-lines-custom-actions.test.ts
@@ -1,6 +1,6 @@
 import assert from 'node:assert/strict';
 import { test } from 'vitest';
-import { formatSnapshotLine } from '../../snapshot/snapshot-lines.ts';
+import { formatSnapshotLine } from '../snapshot-lines.ts';
 
 const mergedCard = {
   ref: 'e72',
